Update ResultsPage Add score!
