+++ frontend/src/game/routes/components/ResultsPage.scss
@@ -1,0 +1,21 @@
+.results {
+  .game-language {
+    margin-bottom: 2rem;
+    .language-info {
+      margin: 0 0 0.5rem 0;
+    }
+
+    .language-selector {
+      position: relative;
+      justify-content: center;
+      top: initial;
+      right: initial;
+
+      padding-bottom: 0.25rem;
+    }
+  }
+
+  .button-row {
+    margin-bottom: 4rem;
+  }
+}
